updated test cases, token verification for creating/deleting blogs

+++ controllers/blogs.js
@@ -1,8 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
-const config = require('../utils/config')
+const userExtractor = require('../utils/user_extractor.js')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user')
@@ -15,36 +13,39 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 // Create a new blog
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
   // If title or url is missing, 400 bad request
   if (!body.title || ! body.url) {
     return response.status(400).json({ error: 'missing title/url property' })
   }
-  console.log(request.token)
-  const decodedToken = jwt.verify(request.token, config.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'invalid token' })
-  }
-  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title,
-    author: user.name,
+    author: request.user.name,
     url: body.url,
     likes: body.likes ? body.likes : 0,
-    user: user._id
+    user: request.user._id
   })
   let savedBlog = await blog.save()
-  user.blogs = user.blogs.concat(savedBlog.id)
-  await user.save()
+  request.user.blogs = request.user.blogs.concat(savedBlog.id)
+  await request.user.save()
 
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+// Delete a blog
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
   const id = request.params.id
+  const blog = await Blog.findById(id)
+  // Unauthorized
+  if (blog.user.toString() !== request.user._id.toString()) {
+    return response.status(401).json({ error: 'unauthorized to delete this blog' })
+  }
   await Blog.findByIdAndDelete(id)
+  request.user.blogs = request.user.blogs.filter(blog => blog._id.toString() !== id)
+  await request.user.save()
+
   response.status(204).end()
 })
